[ts-http-runtime] add azure diff check tool (#31285)
### Packages impacted by this PR

- `@typespec/ts-http-runtime`

### Describe the problem that is addressed by this PR

We have had issues with the unbranded Core package becoming out of sync
with the Azure packages. This PR adds a report which is a diff between
the Azure Core packages and the unbranded package and causes the lint
step to fail if there are any differences between Azure and unbranded
that do not appear in the report. Running the `lint:fix` script will
update the diff report. This will mean that:

- Making an Azure Core change without a corresponding change to
unbranded package, either in the source code or in the diff report, will
result in a CI failure
- Changes to the diff report can be reviewed so that any differences
between Azure and unbranded implementations of a feature can be easily
compared

Why use the lint script (and lint:fix)? This is already run by CI, so by
using it we don't need to make any changes to pipeline steps to get this
working. This script is a one-off and so I just want to do the bare
minimum to get things working 😄