[core-client-rest] use helpers for isCredential (#31494)
### Packages impacted by this PR

- `@azure-rest/core-client`

### Describe the problem that is addressed by this PR

Simplify `isCredential` helper used in getClient.ts. This change was
already made in the unbranded package but not in Azure core.

No change in the Unbranded diff report for this PR because I made the
change and updated the diff as part of #31487, just forgot to check in
this particular change. The unbranded diff check is currently reporting
a failure on main and this change should fix that.