Use federated auth for smoke testing (#31498)
The smoke-tests pipeline has been disabled but I still cleaned up to
smoke tests to remove usages of sub-config-azure-cloud-test-resources
and try to use a service connection. When we decide to re-enable the
smoke tests pipeline we will want to update the deployment to use
federated auth as well.

@xirzec you disabled this pipeline 1/10/2023. Do you have any plans to
try and re-enable it? If not perhaps we just delete this instead.