`js - pullrequest` (#31420)
This pullrequest adds a definition for an expanding and contracting
build that will be used to replace all `js - <service> - ci` build
definitions in the `public` devops project. This is an overall package
selection change, where we find the targeted packages using
`Save-Package-Properties.ps1` primarily. Analyze checks have also been
updated to honor this new selection methodology, rather than the hardcoded
artifactslist.

Only `js - pullrequest` will actually honor the expand and contract for
the packages that were changed in the PR. Meanwhile the `internal`
versions of the `service`-specific definitions will still exist, still
select packages for the service directory, and still be used to release
as normal.

`js - pullrequest` will run _alongside_ the existing builds for a short
period while we identify any outstanding bugs or collection adjustments.
After we're happy, we'll disable all public js builds in favor of the
single `js - pullrequest` build.

---------

Co-authored-by: Ben Broderick Phillips <bebroder@microsoft.com>