[ServiceBus] fix issue of accepting session with empty string name
We have a check for asking for any session, i.e., not providing a session id),
but not receiving any. However, the check is `if (!this._providedSessionId &&
!receivedSessionId)` which also evaluated to `true` for session id of empty
string, which is valid.

This PR changes it to check for `undefined` so that session id of empty string
doesn't lead to error being thrown.