Update Sinon Library Version for Chat JS SDK (#31737)
This PR is used to update the Sinon library version for Chat JS SDK.

**Why is this change needed?**
We updated the Sinon Library Version of Signaling Client Code Repo to
resolve a government alert. And Chat SDK will depend on signaling
client. If they have the same dependency installed, the dependency
version should be the same.
As a result, we are supposed to update the library version of JS SDK for
version number alignment.

**Testing**
1.Unit tests passed after the change
2.E2E test passed using the sample app