fix(playwrighttesting): Add X-MS-Package-Version header to connect session API (#31745)
### Packages impacted by this PR
@azure/microsoft-playwright-testing

### Issues associated with this PR
Added X-MS-Package-Version header to connect session API for language
and package version decoding.

### Need to discuss
The getReporterVersion function returns the reporter version, which
includes a special character ("."). This caused tests to fail with the
error: "Invalid character in header content ["x-ms-package-version"]."
Therefore, the characters were escaped.
Since we are passing the header from the package, this needs to be
handled by the server. We are encoding it here, but it needs to be
decoded on the server side.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_
yes

### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)