Debug configuration for vitest (#31752)
### Packages impacted by this PR

None

### Issues associated with this PR

None

### Describe the problem that is addressed by this PR

Using the JavaScript Debug Terminal and auto-attaching to every
subprocess results in a long wait time to hit breakpoints when debugging. To
improve the experience, we can include a launch.json configuration that allows
running vitest tests and only attaching to the vitest process itself.

Because we need the test proxy running, using dev-tool is the
recommended way to run tests. This PR adds the proper configuration 
to support that scenario.

While here I also went ahead and removed `node_modules` from VSCode's
exclusion list so they show up in the file explorer. Happy to revert that but I
find it difficult to dig into dependencies and set breakpoints in node modules 
when they are hidden.