[communication] Move @azure/communication-call-automation to ESM/vitest (#31680)
### Packages impacted by this PR

- @azure/communication-call-automation

### Issues associated with this PR

- https://github.com/Azure/azure-sdk-for-js/issues/31338

### Describe the problem that is addressed by this PR

Migrates package to ESM/vitest. Note that ServiceBus has issues with
Browser loads that we can fix in a later revision as Buffer is not
defined when you try to import any files. But then including Buffer then
also causes issues with the unit tests.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)