[OpenAI] Precise Typechecking (#31786)
Behind the scene, TSHY rewrites tsconfig.json to build the source code
of the library but it ignores tests and samples. This PR breaks the
monolith tsconfig.json into three ones, one for building the source, one
for the tests, and one for the samples. It utilizes [Project
References](https://www.typescriptlang.org/docs/handbook/project-references.html)
to create a TS project each. This setup makes sure type errors are
adequately caught in each of those projects.

Additionally, the PR disables rolling up d.ts files and fixes a type
error in tests.