[ServiceBus] fix issue of accepting session with empty string name (#31736)
We have a check for when asking for any session, i.e., not providing a
session id), but not receiving a link having a session id. However, the
check is `if (!this._providedSessionId && !receivedSessionId)` which
also evaluates to `true` for session id of empty string, which is valid
session id.

This PR changes it to check for `undefined` so that session id of empty
string doesn't lead to error being thrown.