[core] bind challengeCallbacks (#31961)
### Packages impacted by this PR

- `@azure/core-rest-pipeline`
- `@azure/container-registry`

### Issues associated with this PR

- Fix #31959 
- Fix #31934

### Describe the problem that is addressed by this PR

Container Registry's custom challenge handler [is implemented as a
class](https://github.com/Azure/azure-sdk-for-js/blob/main/sdk/containerregistry/container-registry/src/containerRegistryChallengeHandler.ts#L38).
Its `authorizeRequest` and `authorizeRequestOnChallenge` use class
fields to store the credential and token cycler. However, in the change
to enable CAE support (#31501), we assign the class methods to a new
object without binding them to the original object, meaning that the
fields are no longer accessible using `this`.

This PR fixes the issue by binding the methods to the original object.

### Are there test cases added in this PR? _(If not, why?)_

Yes, added a test case.

### Provide a list of related PRs _(if any)_

- #31501