Update docIndex to kick off the docs SDK CI run for main (#32083)
Right now, the docs SDK CI runs run a scheduled run against main and the
timing of the runs is isn't great. The solution here is to have the
docIndex runs kick off SDK CI the run against main the same way we're
kicking off the run against the daily branch. This will remove anything
fiddly.

As part of this change, add a parameter, defaulting to false, to
determine whether or not the SDK CI run should be kicked off.

This is what the parameter I'd added looks like. It's the second
checkbox down, unchecked by default.

![image](https://github.com/user-attachments/assets/5b2ddcbb-4b72-46a2-a865-35149d94b4bc)


This is the [docIndex run with it
unchecked](https://dev.azure.com/azure-sdk/internal/_build/results?buildId=4391936&view=results).
Notice it didn't kick off the SDK CI run.
This is the [docIndex run with it
checked](https://dev.azure.com/azure-sdk/internal/_build/results?buildId=4391959&view=results).
Notice the SDK CI run was kicked off.