[EngSys] remove dev dependency `decompress`
and replace usage with `tar` and `unzipper`.

`decompress` pulls in `decompress-tar` which has a vulnerability
https://security.snyk.io/package/npm/decompress-tar. However, `decompress-tar`
hasn't been updated for seven years, and there's no sign of this issue being
addressed in the future.

This PR switches to use more actively maintained packages `tar` and `unzipper`
to extract from compressed archives and removes `decompress` dependency.