[ConfidentialLedger] Generate SDK based on preview-2024-01-26 spec (#31987)
### Packages impacted by this PR
@azure-rest/confidential-ledger

### Issues associated with this PR
Add a single API - the ability to *list* users from the "app/users"
endpoint. Please see the
[doc](https://microsoft-my.sharepoint-df.com/:w:/g/personal/tachou_microsoft_com1/EVUirfTSrlhJslGy0i_eCOgBP_wb0gDTlBZOBynm5wWoCw?e=oaZcv4)
here.

### Describe the problem that is addressed by this PR
Customers can `GET` details of a user from the `app/users/{userId}`
endpoint, but needed to know the `userId` beforehand. The API was
amended in `preview-2024-01-26` to include the ability to list all users
from the `app/users` endpoint.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_
Yes

### Provide a list of related PRs _(if any)_
[Spec API
Change](https://github.com/Azure/azure-rest-api-specs/pull/27564)

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_


### Checklists
- [X] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [X] Added a changelog (if necessary)