[core amqp] Fix build warnings (#32252)
### Packages impacted by this PR
@azure/core-amqp

### Issues associated with this PR
https://github.com/Azure/azure-sdk-for-js/issues/32227

### Describe the problem that is addressed by this PR
The library heavily relies on manually written definitions of web types
that conflict with TS's DOM types. This PR cleans all those types and
simply includes DOM in the typing context.

This issue of the API extractor warnings started to occur after
https://github.com/Azure/azure-sdk-for-js/commit/68b3c7bbb7ec48cc3fd4cec1a35e528ee053059c
has been merged. The root cause of the issue is that the older version
of the `tsconfig.json` file was extending from the root one which sets
the `lib` property to an empty list so `DOM` wasn't included in the
typing context for the API extractor. However,
https://github.com/Azure/azure-sdk-for-js/commit/68b3c7bbb7ec48cc3fd4cec1a35e528ee053059c
removes this `extends` clause which brings `lib` value to be the default
one which includes `DOM`.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
Alternatively, we could still set the `lib` property to the empty list
for the API extractor to not warn about those conflicts. However, I fail
to see the advantage of keeping manually written versions of web types
instead of just importing DOM. Such shims should be provided only if web
types are part of the public API itself.

### Are there test cases added in this PR? _(If not, why?)_
N/A

### Provide a list of related PRs _(if any)_
N/A

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)