Refactor error handling with OpenAIError class (#32284)
### Packages impacted by this PR


### Issues associated with this PR


### Describe the problem that is addressed by this PR
Improve error management by replacing the existing error handling with
the OpenAIError class, ensuring better clarity and consistency in error
reporting. Additionally, address an issue related to setting required
actions on ThreadRunOutput.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)