[engsys] Fix rush-runner (#32392)
### Packages impacted by this PR

All packages using CI

### Describe the problem that is addressed by this PR

@qiaozha noticed that tests were no longer failing the build. After some
investigation, I realized that PR #32148 introduced a regression with
how the process exitCode was being propagated. Previously, we were
setting it globally via `process.exitCode` where we spawn the node
process, but this was refactored to use return codes.

However, the return code was never used when the task was anything
except `lint` and `check-format` meaning that build and test operations
did not report failure correctly.

In the future, we should figure out a better way to abstract and test
this script's return codes, but I'd like to get a fix in to prevent bad
PRs from being merged in the meantime.