[@azure/monitor-query] Skip failing test in the Monitor Query SDK (#32419)
### Packages impacted by this PR

`@azure/monitor-query`

### Issues associated with this PR

https://github.com/Azure/azure-sdk-for-js/issues/32413

### Describe the problem that is addressed by this PR

An issue https://github.com/Azure/azure-sdk-for-js/issues/32413 has been
created to address the live tests pipeline failure for Monitor Query
SDK. On analysis, I have found that there is only one test failing in
the pipeline i.e. `LogsQueryClient live tests - server timeout` in
`logsQueryClient.spec.ts` file.

Now, if I look into this test, here is the logic:
1. Execute a query `range x from 1 to ${randomLimit} step 1 | count`
with the `serverTimeoutInSeconds` set to 1.
2. This query is supposed to encounter a timeout error and be thrown.
(If you look at the recordings in playback mode, you can confirm that
the error is thrown).
3. But, now with the live mode the error is not thrown. As a result, the
test is failing.

So, I wanted to discuss this with the service team to understand what
has changed and probably update the query to ensure that the error is
thrown. But, until that I do not want the live tests pipeline to be
blocked. So, I am skipping this test for now. But, I will not be closing
the original issue until I put in a permanent fix for the same.

With the skipped test, I was able to run the live tests and confirm it
is working fine. Here is the link:
https://dev.azure.com/azure-sdk/internal/_build/results?buildId=4450540&view=results

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

No special design considerations.

### Are there test cases added in this PR? _(If not, why?)_

No. The error itself is happening in a test case.

### Provide a list of related PRs _(if any)_

None

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

None

### Checklists
- [X] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)