update CHANGELOG and samples to reflect additional type renamings in document intelligence (#32327)
<!-- Give your Pull Request a brief description, and answer the
following questions. This helps us find the best reviewers for the PR.
-->

### Packages impacted by this PR

@azure-rest/ai-document-intelligence-rest

### Issues associated with this PR

#32326 

### Describe the problem that is addressed by this PR

Some types renamed in version 1.0.0 are not reflected everywhere

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

 - 

### Are there test cases added in this PR? _(If not, why?)_

This PR only change types, there is no test for that as far as I know.

### Provide a list of related PRs _(if any)_

 -

### Command used to generate this PR: _(Applicable only to SDK release
request PRs)_

- 

### Checklists
- [x] Added impacted package name to the issue description.
- [ ] Does this PR need any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here.)_
- [x] Added a changelog (if necessary).