[engsys] Modernize rush runner (#32444)
### Packages impacted by this PR

None directly, this is a CI script

### Describe the problem that is addressed by this PR

This PR addresses a minor regression where only the last exit code will
be used when linting or formatting. Mostly though it tries to clean up
the code to be easier to reason about by splitting it into more modules,
getting the tests to actually pass, and adding additional tests for exit
codes and invocations.