[core-amqp] set `port` when it is not undefined in connection config
PR #32406 adds the `port` property to parsed connection configuration even it is
`undefined`. While the effect is similar, it broke a Event Hub unit test that
expects the `port` property to not exist when a connection string does not
specify a port.

This PR changes to only add the `port` property when it is not `undefined`.