[dev-tool] use package scope when generating sample apiref link
Currently the apiref link template has a hardcoded `@azure` scope so it doesn't
work for packages with `@azure-rest` scope. Those package relies on the
customized `apiRef` fields in the samples configuration section of package.json.
This PR adds package scope variable to the apiref link template as well so the
link generatation works for both `@azure` and `@azure-rest` scopes.  These two
scopes are the only ones that have SDK API reference pages currently.