[dev-tool] Generate samples tsconfig based on tsconfig.samples.json (#32474)
`samples/<version>/typescript/tsconfig.json` is being created from a
static definition in dev-tool, disregarding the configs used in
typechecking the `samples-dev` folder. This PR fixes this issue by using
the `resolveConfig` abstraction to create a resolved version of
`tsconfig.samples.json` and use that in the typescript samples folder.
Here is an example:
```json
{
  "compilerOptions": {
    "target": "ES2023",
    "module": "commonjs",
    "lib": [],
    "importHelpers": true,
    "strict": true,
    "alwaysStrict": true,
    "noUnusedLocals": true,
    "noUnusedParameters": true,
    "noImplicitReturns": true,
    "noFallthroughCasesInSwitch": true,
    "forceConsistentCasingInFileNames": true,
    "moduleResolution": "node10",
    "allowSyntheticDefaultImports": true,
    "esModuleInterop": true,
    "outDir": "./dist",
    "resolveJsonModule": true
  },
  "include": [
    "./src"
  ]
}
```

The PR preserves `commonjs` as the target module of the compiled code
but I wonder if we want to move to esm? We can leave that discussion to
a future PR.

Also, this PR removes some of the outdated syntax restrictions related
to older NodeJS versions.