[dev-tool] use package scope when generating sample apiref link (#32465)
Currently the apiref link template has a hardcoded `@azure` scope so it doesn't
work for packages with `@azure-rest` scope. Those packages rely on the
customized `apiRefLink` fields in the sample's configuration section of
package.json. This PR adds package scope variable to the apiref link template as
well so the link generation works for both `@azure` and `@azure-rest` scopes.
These two scopes are the only ones that have SDK API reference pages currently.

With this change, customized `apiRefLink` for -rest packages are also removed.