[core-client-rest] correctly throw RestError when response has no body (#32566)
### Packages impacted by this PR

- `@azure-rest/core-client`
- `@typespec/ts-http-runtime`

### Describe the problem that is addressed by this PR

Noticed with Jose when testing the todo end-to-end demo TypeSpec.

If we receive an unexpected status from the service (an error or
otherwise), we try to create an error with `createRestError`. It is
possible that a service might for some reason return an error response
with no body, intentionally or otherwise. Currently this causes a
TypeError when createRestError tries to read properties off of the body.
This PR guards against the body being undefined so that the expected
`RestError` is thrown instead.