[eslint-plugin] update regex of matching TypeScript file extentions
Some of our custom rules perform operations on TypeScript files only and they
use a regex of `/\.ts$/.test(fileName)` to check whether a file is TypeScript
code file or not. With the migration to ESM we now can have .mts and .cts files
as well. This PR updates the pattern to also match those files.