[opentelemetry] skip marking 304 RestError as an error on spans (#32666)
### Packages impacted by this PR

@azure/opentelemetry-instrumentation-azure-sdk

### Issues associated with this PR

#32454

### Describe the problem that is addressed by this PR

When a conditional request comes back from the server with a status of
304, that signals that the cached resource is still valid. While some libraries
handle that gracefully, others may throw an error due to backwards
compatibility.

Regardless of whether an error is thrown to the user or not, a 304 is
generally not an error from a tracing perspective and can be suppressed.

In passing, I am also updating the implementation to skip setting span
status to OK in general scenarios as per the OTel spec.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

https://github.com/Azure/azure-sdk-for-js/pull/32663 explores doing this
at the core-tracing level; however, I think these semantics are specific to
OTel and this approach allows core-tracing to remain simple.