Added support for Index Metrics V2 (#32389)
### Packages impacted by this PR
@azure/cosmos

### Issues associated with this PR
#32064 

### Describe the problem that is addressed by this PR
The backend currently returns Index Metrics in string format when
populateIndexMetrics is sent true in query options. The backend has now
introduced a V2 version of index metrics that returns the response in
JSON format. This PR implements the changes for SDK to handle the new
JSON format seamlessly. This change ensures compatibility with the
updated backend while maintaining functionality with the previous
version.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)